feat: 增加地址信息校验[地址间距有的地址不可以直接删除,新增修改按照名称和分类进行唯一校验,显示对应的id和名字]，重载基类返回错误信息

+++ order-manage-admin/src/main/java/com/ywq/address/service/impl/AddressInfoServiceImpl.java
@@ -1,11 +1,15 @@
 package com.ywq.address.service.impl;
 
 import java.util.List;
-import org.springframework.beans.factory.annotation.Autowired;
+
+import com.ywq.address.domain.AddressDistanceInfo;
+import com.ywq.common.utils.StringUtils;
 import org.springframework.stereotype.Service;
 import com.ywq.address.mapper.AddressInfoMapper;
 import com.ywq.address.domain.AddressInfo;
 import com.ywq.address.service.IAddressInfoService;
+
+import javax.annotation.Resource;
 
 /**
  * 地址经纬度信息Service业务层处理
@@ -16,8 +20,11 @@ import com.ywq.address.service.IAddressInfoService;
 @Service
 public class AddressInfoServiceImpl implements IAddressInfoService 
 {
-    @Autowired
+    @Resource
     private AddressInfoMapper addressInfoMapper;
+
+    @Resource
+    private AddressDistanceInfoServiceImpl addressDistanceInfoService;
 
     /**
      * 查询地址经纬度信息
@@ -44,6 +51,18 @@ public class AddressInfoServiceImpl implements IAddressInfoService
     }
 
     /**
+     * 查询地址经纬度信息数量
+     *
+     * @param addressInfo
+     * @return
+     */
+    @Override
+    public int selectAddressInfoCount(AddressInfo addressInfo)
+    {
+        return addressInfoMapper.selectAddressInfoCount(addressInfo);
+    }
+
+    /**
      * 新增地址经纬度信息
      * 
      * @param addressInfo 地址经纬度信息
@@ -52,7 +71,18 @@ public class AddressInfoServiceImpl implements IAddressInfoService
     @Override
     public int insertAddressInfo(AddressInfo addressInfo)
     {
-        return addressInfoMapper.insertAddressInfo(addressInfo);
+        AddressInfo query = new AddressInfo();
+        query.setName(addressInfo.getName());
+        query.setClassId(addressInfo.getClassId());
+        int addressInfos = this.selectAddressInfoCount(query);
+        if (addressInfos > 0){
+            return -1;
+        }
+        int answer = addressInfoMapper.insertAddressInfo(addressInfo);
+        if (answer != 0){
+            addressInfoMapper.saveAddressCode(addressInfo);
+        }
+        return answer;
     }
 
     /**
@@ -64,6 +94,14 @@ public class AddressInfoServiceImpl implements IAddressInfoService
     @Override
     public int updateAddressInfo(AddressInfo addressInfo)
     {
+        AddressInfo query = new AddressInfo();
+        query.setId(addressInfo.getId());
+        query.setName(addressInfo.getName());
+        query.setClassId(addressInfo.getClassId());
+        int addressInfos = this.selectAddressInfoCount(query);
+        if (addressInfos > 0){
+            return -1;
+        }
         return addressInfoMapper.updateAddressInfo(addressInfo);
     }
 
@@ -76,6 +114,14 @@ public class AddressInfoServiceImpl implements IAddressInfoService
     @Override
     public int deleteAddressInfoByIds(Long[] ids)
     {
+        String idsStr = StringUtils.join(ids,",");
+        AddressDistanceInfo addressDistanceInfo = new AddressDistanceInfo();
+        addressDistanceInfo.setCodeXName(idsStr);
+        addressDistanceInfo.setCodeYName(idsStr);
+        int count = addressDistanceInfoService.selectAddressDistanceInfoCount(addressDistanceInfo);
+        if (count > 0){
+            return -1;
+        }
         return addressInfoMapper.deleteAddressInfoByIds(ids);
     }
 
@@ -88,6 +134,14 @@ public class AddressInfoServiceImpl implements IAddressInfoService
     @Override
     public int deleteAddressInfoById(Long id)
     {
+        String idsStr = id.toString();
+        AddressDistanceInfo addressDistanceInfo = new AddressDistanceInfo();
+        addressDistanceInfo.setCodeXName(idsStr);
+        addressDistanceInfo.setCodeYName(idsStr);
+        int count = addressDistanceInfoService.selectAddressDistanceInfoCount(addressDistanceInfo);
+        if (count > 0){
+            return -1;
+        }
         return addressInfoMapper.deleteAddressInfoById(id);
     }
 }
